Crate sui_light_client

Structs

Args ğŸ”’

A light client for the Sui blockchain

CheckpointsList ğŸ”’

Config ğŸ”’

RemotePackageStore ğŸ”’

Enums

SCommands ğŸ”’

Functions

check_and_sync_checkpoints ğŸ”’

download_checkpoint_summary ğŸ”’

extract_verified_effects_and_events ğŸ”’

get_full_checkpoint ğŸ”’

get_verified_effects_and_events ğŸ”’

get_verified_object ğŸ”’

main

read_checkpoint ğŸ”’

read_checkpoint_general ğŸ”’

read_checkpoint_list ğŸ”’

sync_checkpoint_list_to_latest ğŸ”’

Run binary search to for each end of epoch checkpoint that is missing between the latest on the list and the latest checkpoint.

write_checkpoint ğŸ”’

write_checkpoint_general ğŸ”’

write_checkpoint_list ğŸ”’