Function sui_light_client::construct::construct_proof

pub fn construct_proof(
    targets: ProofTarget,
    data: &CheckpointData,
) -> Result<Proof>

Construct a proof from the given checkpoint data and proof targets.

Only minimal cheaper checks are performed to ensure the proof is valid. If you need guaranteed validity consider calling verify_proof function on the constructed proof. It either returns Ok with a proof, or Err with a description of the error.