Crate sui_light_client

Modules

checkpoint

config

construct

graphql

object_store

package_store

proof

verifier

Structs

Proof

A proof for specific targets. It certifies a checkpoint summary and optionally includes transaction evidence to certify objects and events.

ProofTarget

Define aspect of Sui state that need to be certified in a proof

TransactionProof

Part of a proof that provides evidence relating to a specific transaction to certify objects and events.

Functions

construct_proof

Construct a proof from the given checkpoint data and proof targets.

verify_proof

Verify a proof against a committee. A proof is valid if it certifies the checkpoint summary and optionally includes transaction evidence to certify objects and events.